+++ tests/graph/loaders/test_validator.py
@@ -4,10 +4,7 @@ from cognite.neat.graph.stores import NeatGraphStore
 
 
 def test_orphan_assets(transformation_rules, solution_knowledge_graph):
-    assets = rdf2assets(
-        NeatGraphStore(solution_knowledge_graph),
-        transformation_rules,
-    )
+    assets = rdf2assets(NeatGraphStore(solution_knowledge_graph), transformation_rules, data_set_id=123456)
 
     asset_external_ids = [asset.get("external_id") for asset in assets.values()]
 
@@ -21,10 +18,7 @@ def test_orphan_assets(transformation_rules, solution_knowledge_graph):
 
 
 def test_circular_assets(transformation_rules, solution_knowledge_graph):
-    assets = rdf2assets(
-        NeatGraphStore(solution_knowledge_graph),
-        transformation_rules,
-    )
+    assets = rdf2assets(NeatGraphStore(solution_knowledge_graph), transformation_rules, data_set_id=123456)
 
     # Create a circular asset hierarchy
     assets["f176960e-9aeb-11e5-91da-b8763fd99c5f"]["parent_external_id"] = "2dd9017c-bdfb-11e5-94fa-c8f73332c8f4"